+++ src/main/kotlin/net/ndrei/teslapoweredthingies/machines/fluidcompoundproducer/FluidCompoundProducerRecipe.kt
@@ -1,0 +1,8 @@
+package net.ndrei.teslapoweredthingies.machines.itemcompoundproducer
+
+import net.minecraftforge.fluids.FluidStack
+
+/**
+ * Created by CF on 2017-07-13.
+ */
+class FluidCompoundProducerRecipe(val inputA: FluidStack, val inputB: FluidStack, val output: FluidStack)